# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KnxMonitor.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Projects/upload": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Uploads an ETS project file and starts the import as a background job.",
        "description": "Returns as soon as the job exists — the parse itself runs in the background, so poll GET imports/{id} for progress. Only files ending in .knxproj are accepted, and the request body is capped at 200 MB. If the archive is password protected the job stops in WaitingForInput and lists what it needs; feed those inputs to POST imports/{id}/provide-input, then the job continues on its own. Re-uploading a project that was imported before is recognised by its ETS project id and updates the existing project in place, so recorded telegrams keep their group-address link. A genuinely new project is auto-activated when no other project is active yet.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The created import job: id, status, the step list and (once finished) the project id and counts."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects/imports/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Returns the current state of an import job.",
        "description": "Jobs are held in memory only. A job that has completed, failed or been cancelled is swept an hour after it terminated (the cleanup worker runs every five minutes) and is reported as not found from then on.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job id returned by the upload call.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status, overall progress, per-step detail and the outstanding requirements."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Marks an import job as cancelled.",
        "description": "This flips the job state only. A parse already running in the background is not interrupted and a project it has written stays. Unknown job ids are accepted without effect.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects/imports/{id}/provide-input": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Supplies one requested input — project password, keyring file or keyring password — to a waiting import job.",
        "description": "Each call fulfils exactly the requirement named in type; the keyring file is passed Base64-encoded. The call is only accepted while the job is in WaitingForInput and rejected otherwise. Supplying the project password triggers a second look inside the now decryptable archive: if KNX Secure devices turn up, an optional keyring file and keyring password requirement are added and the job stays in WaitingForInput. As soon as every listed requirement is marked fulfilled, the import resumes in the background — poll GET imports/{id} from there. The password is not verified at this point; a wrong one surfaces as a failed job.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job id of the waiting import.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvideInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvideInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProvideInputDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Lists all imported projects.",
        "responses": {
          "200": {
            "description": "Per project: name, source file name, import date, active flag and the group-address and device counts."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Returns one project together with its group addresses and devices.",
        "description": "Both lists come back complete and unpaged, so the response grows with the size of the project.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Deletes a project and everything parsed from it.",
        "description": "Removes the project's group addresses, devices, locations, communication objects, group ranges and stored keyring material in one transaction. Recorded telegrams are kept but lose their group-address link and are shown undecoded from then on. Deleting the active project leaves no project active; the bus link stays up.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects/{id}/activate": {
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Makes this project the active one.",
        "description": "Activation is exclusive: any other active project is deactivated in the same transaction, so at most one project is ever active. Afterwards the group-address cache is refreshed and incoming telegrams resolve to this project's names and DPTs. The bus connection is deliberately left untouched — activating decodes, it does not connect.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects/{id}/deactivate": {
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Deactivates the project, leaving no project active.",
        "description": "The group-address cache is cleared, so telegrams are shown undecoded — raw hex, without name or DPT. Monitoring keeps running: the bus link stays up. Deactivating a project that is already inactive succeeds and changes nothing.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects/{id}/delete-preview": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Reports what deleting this project would affect, for the confirmation dialog.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The project's group-address and device counts plus the number of recorded telegrams mapped to it. Those telegrams survive the deletion — they only lose their group-address mapping."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects/{id}/locations": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Returns the project's locations (buildings, floors, rooms, …).",
        "description": "A flat list, not a tree — the hierarchy is expressed through each entry's parent id. Every entry also carries the device and group addresses assigned to that location.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects/{id}/commobjects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Returns the project's communication objects, optionally only those linked to one group address.",
        "description": "Each entry is enriched with the name, manufacturer and product name of the device behind its physical address, as far as that device is part of the project. Sorted by device address, then by object number.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project id.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "address",
            "in": "query",
            "description": "Group address (m/m/s). When set, only objects whose group-address link matches it exactly are returned.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects/{id}/groupranges": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Returns the project's group ranges — the named main and middle groups — ordered by range start.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Projects/{id}/device": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Resolves a single device of the project by its physical address.",
        "description": "The address is matched exactly, and an unknown project id is reported the same way as an unknown device.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project id.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "address",
            "in": "query",
            "description": "Physical address (a.l.d) to look up. Required.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "ProvideInputDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/RequirementType"
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "keyringFile": {
            "type": "string",
            "nullable": true
          },
          "keyringPassword": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "RequirementType": {
        "type": "integer"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access-Token aus POST /api/Auth/login (ohne \"Bearer \"-Präfix einfügen).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}